main.asm
A combined organstop and metronome. The organstop can make the standard sound 440 Hz "la". The metronome can read in the speed value and use the microcomputer loudspeaker to send out a 4/4 "dash" sound in this speed value. The total beat number and the serial beat number of the metronome can be displayed on the seven-digital-tube.